web/requirements.txt
google-cloud-firestore==0.31.0
Flask==1.0.2
google-auth==1.6.3
Jinja2==2.10